+++ src/Phan/Analysis.php
@@ -108,11 +108,7 @@ class Analysis
         }
         try {
             $node = Parser::parseCode($code_base, $context, $request, $file_path, $file_contents, $suppress_parse_errors);
-        } catch (ParseError $_) {
-            return $context;
-        } catch (CompileError $_) {
-            return $context;
-        } catch (ParseException $_) {
+        } catch (ParseError|CompileError|ParseException $_) {
             return $context;
         }
 
@@ -533,13 +529,7 @@ class Analysis
                 return $context;
             }
             $node = Parser::parseCode($code_base, $context, $request, $file_path, $file_contents, false);
-        } catch (ParseException $_) {
-            // Issue::SyntaxError was already emitted.
-            return $context;
-        } catch (ParseError $_) {
-            // Issue::SyntaxError was already emitted.
-            return $context;
-        } catch (CompileError $_) {
+        } catch (ParseException|ParseError|CompileError $_) {
             // Issue::SyntaxError was already emitted.
             return $context;
         }